Fixed RMR calculation.
Note that we consolidate the RMR values with the measurement time of the
full length measurement data frame by slicing the RMR values, do not pad
the RMR values as this increases the RMR artificially. Time points we
report are scaled by the measurement interval length, which is correct.

Note too, that the RMR traces might be a bit shorter, since we use an
averaging window and might thus loose some data points.